+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Changes
 
-* `WebDriver#wait_until_element_disappear` has option of timeout
-* `WebDriver#wait_until_element_disappear` raise `TimeOutError`
-  instead of `RuntimeError`
+* `WebDriver#wait_until_element_disappear`, `WebDriver#wait_until_element_present`
+   has option of timeout
+* `WebDriver#wait_until_element_disappear`, `WebDriver#wait_until_element_present`
+   raise `TimeOutError` instead of `RuntimeError`
 
 ## 0.3.1 (2020-06-03)